fix: compaction
